+++ R/ggplot_stars_fun.R
@@ -118,8 +118,8 @@ ggplot_stars <- function(data,
                          subtitle = NULL,
                          caption = "",
                          font_family = "Helvetica",
-                         font_size_title = 11,
-                         font_size_body = 10,
+                         font_size_title = NULL,
+                         font_size_body = NULL,
                          wrap_title = 70,
                          wrap_subtitle = 80,
                          wrap_caption = 80,
@@ -128,6 +128,15 @@ ggplot_stars <- function(data,
     stop("Please use an stars object as data input")
   if (is.na(sf::st_crs(data)))
     stop("Please assign a coordinate reference system")
+  
+  if(is.null(font_size_title)){
+    if (isMobile == FALSE) font_size_title <- 11
+    else if (isMobile == TRUE) font_size_title <- 15
+  }
+  if(is.null(font_size_body)){
+    if (isMobile == FALSE) font_size_body <- 10
+    else if (isMobile == TRUE) font_size_body <- 14
+  }
   
   plot <- ggplot() +
     theme_stars(
@@ -230,6 +239,7 @@ ggplot_stars <- function(data,
 #' @export
 #' @examples
 #' ggplot_stars_col(data = example_stars_nz_no3n, coastline = nz,
+#'    col_method = "quantile", quantile_cuts = c(0, 0.05, 0.25, 0.5, 0.75, 0.95, 1),
 #'    title = "River modelled median nitrate-nitrogen concentrations, 2013-17")
 ggplot_stars_col <- function(data,
                              col_method = "quantile",
@@ -248,8 +258,8 @@ ggplot_stars_col <- function(data,
                              caption = "",
                              legend_labels = NULL,
                              font_family = "Helvetica",
-                             font_size_title = 11,
-                             font_size_body = 10,
+                             font_size_title = NULL,
+                             font_size_body = NULL,
                              wrap_title = 70,
                              wrap_subtitle = 80,
                              wrap_col_title = 25,
@@ -259,6 +269,15 @@ ggplot_stars_col <- function(data,
     stop("Please use an stars object as data input")
   if (is.na(sf::st_crs(data)))
     stop("Please assign a coordinate reference system")
+  
+  if(is.null(font_size_title)){
+    if (isMobile == FALSE) font_size_title <- 11
+    else if (isMobile == TRUE) font_size_title <- 15
+  }
+  if(is.null(font_size_body)){
+    if (isMobile == FALSE) font_size_body <- 10
+    else if (isMobile == TRUE) font_size_body <- 14
+  }
   
   plot <- ggplot() +
     theme_stars(
@@ -472,8 +491,8 @@ ggplot_stars_facet <- function(data,
                                subtitle = NULL,
                                caption = "",
                                font_family = "Helvetica",
-                               font_size_title = 11,
-                               font_size_body = 10,
+                               font_size_title = NULL,
+                               font_size_body = NULL,
                                wrap_title = 70,
                                wrap_subtitle = 80,
                                wrap_caption = 80,
@@ -482,6 +501,15 @@ ggplot_stars_facet <- function(data,
     stop("Please use an stars object as data input")
   if (is.na(sf::st_crs(data)))
     stop("Please assign a coordinate reference system")
+  
+  if(is.null(font_size_title)){
+    if (isMobile == FALSE) font_size_title <- 11
+    else if (isMobile == TRUE) font_size_title <- 15
+  }
+  if(is.null(font_size_body)){
+    if (isMobile == FALSE) font_size_body <- 10
+    else if (isMobile == TRUE) font_size_body <- 14
+  }
   
   plot <- ggplot() +
     theme_stars(
@@ -622,7 +650,8 @@ ggplot_stars_facet <- function(data,
 #' map_data <- c(map_data1, map_data2)
 #'
 #' ggplot_stars_col_facet(data = map_data, coastline = nz,
-#'      title = "River modelled nutrient concentrations, 2013-17")
+#'    col_method = "quantile", quantile_cuts = c(0, 0.05, 0.25, 0.5, 0.75, 0.95, 1),
+#'    title = "River modelled nutrient concentrations, 2013-17")
 ggplot_stars_col_facet <- function(data,
                                    col_method = "quantile",
                                    quantile_cuts = c(0, 0.25, 0.5, 0.75, 1),
@@ -642,8 +671,8 @@ ggplot_stars_col_facet <- function(data,
                                    caption = "",
                                    legend_labels = NULL,
                                    font_family = "Helvetica",
-                                   font_size_title = 11,
-                                   font_size_body = 10,
+                                   font_size_title = NULL,
+                                   font_size_body = NULL,
                                    wrap_title = 70,
                                    wrap_subtitle = 80,
                                    wrap_col_title = 25,
@@ -653,6 +682,15 @@ ggplot_stars_col_facet <- function(data,
     stop("Please use an stars object as data input")
   if (is.na(sf::st_crs(data)))
     stop("Please assign a coordinate reference system")
+  
+  if(is.null(font_size_title)){
+    if (isMobile == FALSE) font_size_title <- 11
+    else if (isMobile == TRUE) font_size_title <- 15
+  }
+  if(is.null(font_size_body)){
+    if (isMobile == FALSE) font_size_body <- 10
+    else if (isMobile == TRUE) font_size_body <- 14
+  }
   
   plot <- ggplot() +
     theme_stars(