+++ R/ggplot_vbar_fun.R
@@ -161,8 +161,8 @@ ggplot_vbar <- function(data,
                         y_title = "[Y title]",
                         caption = "",
                         font_family = "Helvetica",
-                        font_size_title = 11,
-                        font_size_body = 10,
+                        font_size_title = NULL,
+                        font_size_body = NULL,
                         wrap_title = 70,
                         wrap_subtitle = 80,
                         wrap_x_title = 50,
@@ -181,11 +181,20 @@ ggplot_vbar <- function(data,
   if (!(is.numeric(x_var_vector) | lubridate::is.Date(x_var_vector))) stop("Please use a numeric or date x variable for a vertical bar plot. ")
   if (!is.numeric(y_var_vector)) stop("Please use a numeric y variable for a vertical bar plot")
   
-  if(min(y_var_vector) < 0 & y_scale_zero == TRUE) {
+  if(min(y_var_vector, na.rm = TRUE) < 0 & y_scale_zero == TRUE) {
     y_scale_zero <- FALSE
     message("y_scale_zero must be FALSE as data contains values less than zero")
   }
   
+  if(is.null(font_size_title)){
+    if (isMobile == FALSE) font_size_title <- 11
+    else if (isMobile == TRUE) font_size_title <- 15
+  }
+  if(is.null(font_size_body)){
+    if (isMobile == FALSE) font_size_body <- 10
+    else if (isMobile == TRUE) font_size_body <- 14
+  }
+
   if (is.null(pal)) pal <- pal_snz
   
   plot <- ggplot(data, aes(x = !!x_var, y = !!y_var)) +
@@ -266,7 +275,7 @@ ggplot_vbar <- function(data,
   if (lubridate::is.Date(x_var_vector)) {
     plot <- plot +
       scale_x_date(
-        expand = c(0.05, 0),
+        expand = c(0, 0),
         breaks = x_scale_breaks,
         limits = x_scale_limits,
         labels = scales::date_format(x_scale_date_format)
@@ -274,7 +283,7 @@ ggplot_vbar <- function(data,
   }
   else if (is.numeric(x_var_vector)) {
     plot <- plot +
-      scale_x_continuous(expand = c(0.05, 0),
+      scale_x_continuous(expand = c(0, 0),
                          breaks = x_scale_breaks,
                          limits = x_scale_limits,
                          oob = scales::rescale_none)
@@ -378,8 +387,8 @@ ggplot_vbar_col <-
            caption = "",
            legend_labels = NULL,
            font_family = "Helvetica",
-           font_size_title = 11,
-           font_size_body = 10,
+           font_size_title = NULL,
+           font_size_body = NULL,
            wrap_title = 70,
            wrap_subtitle = 80,
            wrap_x_title = 50,
@@ -402,10 +411,19 @@ ggplot_vbar_col <-
     if (!is.numeric(y_var_vector)) stop("Please use a numeric y variable for a vertical bar plot")
     if (is.numeric(col_var_vector)) stop("Please use a categorical colour variable for a vertical bar plot")
     
+    if(is.null(font_size_title)){
+      if (isMobile == FALSE) font_size_title <- 11
+      else if (isMobile == TRUE) font_size_title <- 15
+    }
+    if(is.null(font_size_body)){
+      if (isMobile == FALSE) font_size_body <- 10
+      else if (isMobile == TRUE) font_size_body <- 14
+    }
+    
     if (position == "stack" & y_scale_trans != "identity") message("simplevis may not perform correctly using a y scale other than identity where position equals stack")
     if (position == "stack" & y_scale_zero == FALSE) message("simplevis may not perform correctly with position equal to stack and y_scale_zero equal to FALSE")
 
-    if(min(y_var_vector) < 0 & y_scale_zero == TRUE) {
+    if(min(y_var_vector, na.rm = TRUE) < 0 & y_scale_zero == TRUE) {
       y_scale_zero <- FALSE
       message("y_scale_zero must be FALSE as data contains values less than zero")
     }
@@ -520,7 +538,7 @@ ggplot_vbar_col <-
     if (lubridate::is.Date(x_var_vector)) {
       plot <- plot +
         scale_x_date(
-          expand = c(0.05, 0),
+          expand = c(0, 0),
           breaks = x_scale_breaks,
           limits = x_scale_limits,
           labels = scales::date_format(x_scale_date_format)
@@ -528,7 +546,7 @@ ggplot_vbar_col <-
     }
     else if (is.numeric(x_var_vector)) {
       plot <- plot +
-        scale_x_continuous(expand = c(0.05, 0),
+        scale_x_continuous(expand = c(0, 0),
                            breaks = x_scale_breaks,
                            limits = x_scale_limits,
                            oob = scales::rescale_none)
@@ -644,8 +662,8 @@ ggplot_vbar_facet <-
            y_title = "[Y title]",
            caption = "",
            font_family = "Helvetica",
-           font_size_title = 11,
-           font_size_body = 10,
+           font_size_title = NULL,
+           font_size_body = NULL,
            wrap_title = 70,
            wrap_subtitle = 80,
            wrap_x_title = 50,
@@ -667,9 +685,18 @@ ggplot_vbar_facet <-
     if (!is.numeric(y_var_vector)) stop("Please use a numeric y variable for a vertical bar plot")
     if (is.numeric(facet_var_vector)) stop("Please use a categorical facet variable for a vertical bar plot")
     
-    if(min(y_var_vector) < 0 & y_scale_zero == TRUE) {
+    if(min(y_var_vector, na.rm = TRUE) < 0 & y_scale_zero == TRUE) {
       y_scale_zero <- FALSE
       message("y_scale_zero must be FALSE as data contains values less than zero")
+    }
+    
+    if(is.null(font_size_title)){
+      if (isMobile == FALSE) font_size_title <- 11
+      else if (isMobile == TRUE) font_size_title <- 15
+    }
+    if(is.null(font_size_body)){
+      if (isMobile == FALSE) font_size_body <- 10
+      else if (isMobile == TRUE) font_size_body <- 14
     }
     
     if (is.null(pal)) pal <- pal_snz
@@ -750,7 +777,7 @@ ggplot_vbar_facet <-
       if (lubridate::is.Date(x_var_vector)) {
         plot <- plot +
           scale_x_date(
-            expand = c(0.05, 0),
+            expand = c(0, 0),
             breaks = x_scale_breaks,
             limits = x_scale_limits,
             labels = scales::date_format(x_scale_date_format)
@@ -758,7 +785,7 @@ ggplot_vbar_facet <-
       }
       else if (is.numeric(x_var_vector)) {
         plot <- plot +
-          scale_x_continuous(expand = c(0.05, 0),
+          scale_x_continuous(expand = c(0, 0),
                              breaks = x_scale_breaks,
                              limits = x_scale_limits,
                              oob = scales::rescale_none)
@@ -901,8 +928,8 @@ ggplot_vbar_col_facet <-
            caption = "",
            legend_labels = NULL,
            font_family = "Helvetica",
-           font_size_title = 11,
-           font_size_body = 10,
+           font_size_title = NULL,
+           font_size_body = NULL,
            wrap_title = 70,
            wrap_subtitle = 80,
            wrap_x_title = 50,
@@ -927,10 +954,19 @@ ggplot_vbar_col_facet <-
     if (!is.numeric(y_var_vector)) stop("Please use a numeric y variable for a vertical bar plot")
     if (is.numeric(facet_var_vector)) stop("Please use a categorical facet variable for a vertical bar plot")
     
+    if(is.null(font_size_title)){
+      if (isMobile == FALSE) font_size_title <- 11
+      else if (isMobile == TRUE) font_size_title <- 15
+    }
+    if(is.null(font_size_body)){
+      if (isMobile == FALSE) font_size_body <- 10
+      else if (isMobile == TRUE) font_size_body <- 14
+    }
+    
     if (position == "stack" & y_scale_trans != "identity") message("simplevis may not perform correctly using a y scale other than identity where position equals stack")
     if (position == "stack" & y_scale_zero == FALSE) message("simplevis may not perform correctly with position equal to stack and y_scale_zero equal to FALSE")
     
-    if(min(y_var_vector) < 0 & y_scale_zero == TRUE) {
+    if(min(y_var_vector, na.rm = TRUE) < 0 & y_scale_zero == TRUE) {
       y_scale_zero <- FALSE
       message("y_scale_zero must be FALSE as data contains values less than zero")
     }
@@ -1042,7 +1078,7 @@ ggplot_vbar_col_facet <-
       if (lubridate::is.Date(x_var_vector)) {
         plot <- plot +
           scale_x_date(
-            expand = c(0.05, 0),
+            expand = c(0, 0),
             breaks = x_scale_breaks,
             limits = x_scale_limits,
             labels = scales::date_format(x_scale_date_format)
@@ -1050,7 +1086,7 @@ ggplot_vbar_col_facet <-
       }
       else if (is.numeric(x_var_vector)) {
         plot <- plot +
-          scale_x_continuous(expand = c(0.05, 0),
+          scale_x_continuous(expand = c(0, 0),
                              breaks = x_scale_breaks,
                              limits = x_scale_limits,
                              oob = scales::rescale_none)