+++ inst/shiny/template1/global.R
@@ -7,3 +7,7 @@ library(simplevis)
 
 # load data
 ### read your RDS data in from data folder here and call it df ###
+df <-  ggplot2::diamonds
+
+color_vector <- sort(unique(df$color))
+