+++ inst/shiny/template2/ui.R
@@ -4,9 +4,8 @@
 app_title <- "Template 2" ### name the app ###
 
 ### function to
-mobileDetect <-
+mobileDetect <- # https://github.com/g3rv4/mobileDetect
   function(inputId, value = 0) {
-    # https://github.com/g3rv4/mobileDetect
     tagList(
       singleton(tags$head(tags$script(src = "js/mobile.js"))),
       tags$input(id = inputId,
@@ -23,35 +22,17 @@ shinyUI(
     tags$head(includeCSS("www/style.css")),
     tags$head(includeScript("www/js/tag-manager.js")),
     mobileDetect("isMobile"),
-    # https://github.com/g3rv4/mobileDetect
-    # tags$head(
-    #   tags$script(src = "https://cdnjs.cloudflare.com/ajax/libs/iframe-resizer/3.5.16/iframeResizer.contentWindow.min.js",
-    #               type = "text/javascript")
-    # ), # https://www.cultureofinsight.com/blog/2018/03/15/2018-03-15-responsive-iframes-for-shiny-apps/
     navbarPage(
       title = HTML("<b>", app_title, "</b>"),
       windowTitle = app_title,
-      tabPanel(
-        "Map",
-        icon = icon("globe-americas", lib = "font-awesome"),
-        sidebarLayout(
-          sidebarPanel(
-            width = 3
-            ### add radioButtons and other widgets in here ###
-          ),
-          mainPanel(width = 9,
-                    fluidRow(
-                      shinycssloaders::withSpinner(leaflet::leafletOutput("map"), type = 7, color = "#A8A8A8")
-                    ))
-        )
-      ),
       tabPanel(
         "Graph",
         icon = icon("chart-area", lib = "font-awesome"),
         sidebarLayout(
           sidebarPanel(
-            width = 3
-            ### add radioButtons and other widgets in here ###
+            width = 3, 
+            # add radioButtons and other widgets in here 
+            radioButtons("plot_color", "Colour", color_vector)
           ),
           mainPanel(
             width = 9,
@@ -75,10 +56,36 @@ shinyUI(
         )
       ),
       tabPanel(
+        "Map",
+        icon = icon("globe-americas", lib = "font-awesome"),
+        sidebarLayout(
+          sidebarPanel(
+            width = 3,
+            # add radioButtons and other widgets in here
+            radioButtons("map_indicator", "Indicator", indicator_vector)
+          ),
+          mainPanel(width = 9,
+                    fluidRow(
+                      shinycssloaders::withSpinner(leaflet::leafletOutput("map"), type = 7, color = "#A8A8A8")
+                    ))
+        )
+      ),
+      tabPanel(
         "Table",
         icon = icon("table", lib = "font-awesome"),
-        fluidRow(
-          shinycssloaders::withSpinner(DT::DTOutput("table"), type = 7, color = "#A8A8A8")
+        # fluidRow(
+        #   shinycssloaders::withSpinner(DT::DTOutput("table"), type = 7, color = "#A8A8A8")
+        # )
+        sidebarLayout( #use this if multiple datasets
+          sidebarPanel(
+            width = 3,
+            ### add radioButtons and other widgets in here ###
+            radioButtons("table_data", "Dataset", c("Diamonds", "Storms"))
+          ),
+          mainPanel(
+            width = 9,
+            shinycssloaders::withSpinner(DT::DTOutput("table"), type = 7, color = "#A8A8A8")
+          )
         )
       ),
       tabPanel(
@@ -92,6 +99,5 @@ shinyUI(
         fluidRow(includeMarkdown("www/about.Rmd"))
       )
     )
-    # HTML('<div data-iframe-height></div>') # https://www.cultureofinsight.com/blog/2018/03/15/2018-03-15-responsive-iframes-for-shiny-apps/
   )
 )